requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.34,<1.0
sentence-transformers>=3.3,<4.0
faiss-cpu>=1.10,<2.0
numpy>=1.26,<3.0
python-dotenv>=1.0,<2.0
requests>=2.32,<3.0
streamlit>=1.41,<2.0
pypdf>=5.4,<6.0
rapidfuzz>=3.9,<4.0
python-multipart>=0.0.9,<1.0
opencv-python>=4.10,<5.0
torch 
transformers
